Refactored native module. Handle nested exceptions. Detect forbidden NTLMv1
