+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.0 - released 2019-10-09
+
+* Detect if incoming challenge is NTLMv1 and the client is not permitted to use NTLMv1
+* Graceful error handling - allows nested errors from the native module
+
 ## 0.2.15 - released 2019-10-04
 
 * Repair npm packaging
