+++ src/win-sso.ts
@@ -14,13 +14,14 @@ try {
  * Creates authentication tokens for NTLM handshake using the executing users credentials.
  */
 export class WinSso {
+  private static NEGOTIATE_NTLM2_KEY = 1<<19;
 
   /**
    * Retrieves the username of the logged in user
    * @returns {string} user name including domain
    */
-  static getUserName(): string {
-    return winSsoAddon.getUserName();
+  static getLogonUserName(): string {
+    return winSsoAddon.getLogonUserName();
   }
 
   private static getChannelBindingsApplicationData(peerCert: PeerCertificate) {
@@ -86,9 +87,32 @@ export class WinSso {
       applicationData = Buffer.alloc(0);
     }
 
-    let token = winSsoAddon.createAuthResponse(inToken, targetHostStr, applicationData);
-    debug('Created NTLM type 3 token', token.toString('base64'));
-    return token;
+    try {
+      let token = winSsoAddon.createAuthResponse(inToken, targetHostStr, applicationData);
+      debug('Created NTLM type 3 token', token.toString('base64'));
+      return token;
+    } catch (err) {
+      if (err.message === 'Could not init security context. Result: -2146893054') {
+        // If incoming token is for NTLMv1, this error can occur when LMCompatibilityLevel prevents the client to send NTLMv1 messages
+        if (WinSso.IsNtlmV1(inToken)) {
+          throw new Error('Could not create NTLM type 3 message. Incoming type 2 message uses NTLMv1, '+
+                          'it is likely that the client is prevented from sending such messages. ' +
+                          'Update target host to use NTLMv2 (recommended) or adjust LMCompatibilityLevel on the client (insecure)');
+        }
+      } else {
+        throw err;
+      }
+    }
+  }
+
+  private static IsNtlmV1(type2message: Buffer): boolean {
+    if (type2message.length >= 24) {
+      const inTokenFlags = type2message.readInt32BE(20);
+      if ((inTokenFlags & WinSso.NEGOTIATE_NTLM2_KEY) === 0) {
+        return true;
+      }
+    }
+    return false;
   }
 
   /**
